Refactored Override and removed ThemeColors
